AutoIT-Bot-Billard
This version is different from the rest. Before I was using 4 bots exactly the same to open simultaneously to make the bot go faster in searching for images but I changed the code in the vb.net BillardBot.exe to open 4 of the same Bot.exe insted of 4 copied versions. Faster for me for when i need to add an image too the code.